% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_expression.R
\name{changelengthofprofile}
\alias{changelengthofprofile}
\title{this is not for the user, only for the author and must be later deleted from here}
\usage{
changelengthofprofile(profilelength)
}
\arguments{
\item{profilelength}{}
}
\description{
this is not for the user, only for the author and must be later deleted from here
}
