% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minpoint.R
\name{minpoint}
\alias{minpoint}
\title{Calculating the lowest points along the track}
\usage{
minpoint(
  raster,
  tracks,
  distance_points_along_line = 1,
  profilelength = 1,
  distancecrossprofilepoints = 0.05
)
}
\arguments{
\item{raster}{a dsm.tif file}

\item{tracks}{a sf file}

\item{distance_points_along_line}{default is 1}
}
\description{
Calculating the lowest points along the track
}
