% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrack.R
\name{readTrack}
\alias{readTrack}
\title{Title}
\usage{
readTrack(
  yourTracksPath = "C:/eagle/trackRcheck/exampledata_noupload/testtrack.gpkg"
)
}
\arguments{
\item{yourTracksPath}{}
}
\description{
Title
}
